backend/requirements.txt
Flask
requests
Pillow
Flask-CORS
google-generativeai